requirements.txt
playwright==1.53.0
python-dotenv==1.1.0
pytest==8.4.1
pytest-playwright==0.7.0
